F.N.D-Project-Codebase/FND-Backend-Pro/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
flask-cors==3.0.10
numpy==1.23.5  # Updated to be compatible with TensorFlow 2.12
pandas==1.5.3  # Downgraded to be compatible with numpy 1.23.5
tensorflow==2.12.0
scikit-learn==1.2.2  # Works with numpy 1.23.5